+++ frontend/src/pages/logged/budget-container/container-panes/pane-transaction/transaction-header/header-add-transaction/transaction-body/components/TransactionBodyInputs.tsx
@@ -1,21 +1,20 @@
 import ModalFormInputGroup from "common/modal-form/components/InputGroup";
 import React from "react";
-import { BoxSeam, CardText, Cash } from "react-bootstrap-icons";
+import { CardText, Cash } from "react-bootstrap-icons";
 import { UseFormRegister } from "react-hook-form";
 
 interface Props {
-  register: UseFormRegister<any>;
+    handleInput: UseFormRegister<any>;
 }
 
-const AddTransactionBody = ({ register }: Props) => {
+const AddTransactionBodyInputs = ({ handleInput }: Props) => {
   const IconCash = (): JSX.Element => <Cash />;
   const IconCard = (): JSX.Element => <CardText />;
-  const IconBox = (): JSX.Element => <BoxSeam />;
 
   return (
     <>
       <ModalFormInputGroup
-        handleInput={register}
+        handleInput={handleInput}
         icon={IconCard}
         formControlProps={{
           placeholder: "Transaction Name",
@@ -25,7 +24,7 @@ const AddTransactionBody = ({ register }: Props) => {
         }}
       />
       <ModalFormInputGroup
-        handleInput={register}
+        handleInput={handleInput}
         icon={IconCash}
         formControlProps={{
           placeholder: "Amount",
@@ -34,18 +33,8 @@ const AddTransactionBody = ({ register }: Props) => {
           name: "amount",
         }}
       />
-      <ModalFormInputGroup
-        handleInput={register}
-        icon={IconBox}
-        formControlProps={{
-          placeholder: "category",
-          type: "text",
-          id: "category",
-          name: "category",
-        }}
-      />
     </>
   );
 };
 
-export default AddTransactionBody;
+export default AddTransactionBodyInputs;