+++ frontend/src/types/expense.d.ts
@@ -1,0 +1,7 @@
+import { Transaction } from "types/transaction";
+
+export type ExpenseCategory = "Food" | "Goods" | "Bills" | "Other";
+
+export type Expense = Transaction & {
+  category: ExpenseCategory;
+};