+++ frontend/src/types/income.d.ts
@@ -1,0 +1,7 @@
+import { Transaction } from "types/transaction";
+
+export type IncomeCategory = "Job" | "Gift" | "Theft" | "Other";
+
+export type Income = Transaction & {
+  category: IncomeCategory;
+};