+++ frontend/src/types/transaction_form.d.ts
@@ -1,0 +1,8 @@
+import { ExpenseCategory } from "types/expense";
+import { IncomeCategory } from "types/income";
+
+export type TransactionForm = {
+  amount: number;
+  name: string;
+  category: ExpenseCategory | IncomeCategory;
+};